completed student table crud
